2018.12.21

+++ db/migrate/20181101070830_create_cuppings.rb
@@ -1,20 +1,20 @@
 class CreateCuppings < ActiveRecord::Migration[5.2]
   def change
     create_table :cuppings do |t|
-      t.integer :clean_cup
-      t.integer :sweet
-      t.integer :acidity
-      t.integer :mouth_feel
-      t.integer :flavor
-      t.integer :after_taste
-      t.integer :balance
-      t.integer :over_all
+      t.float :clean_cup
+      t.float :sweet
+      t.float :acidity
+      t.float :mouth_feel
+      t.float :flavor
+      t.float :after_taste
+      t.float :balance
+      t.float :over_all
       t.text :flavor_coment
       t.string :origin
       t.string :process
       t.string :shop
       t.text :memo
-      t.integer :total_point
+      t.float :total_point
       t.integer :country_id 
 
       t.timestamps
